+++ src/test/suite/unitTests/TreeBuilder.test.ts
@@ -3,76 +3,127 @@ import * as chai from "chai";
 import * as chaiSubset from "chai-subset";
 import { CONSTANTS } from "../../../constants";
 import { TreeBuilder } from "../../../TreeBuilder";
-import { Format, FormatModel, TreeItem } from "../../../types";
+import { DetectedModel, Format, FormatModel, TreeItem } from "../../../types";
 
 suite('TreeBuilder Tests', () => {
     chai.use(chaiSubset);
 
-    suite('generateTreeItemModel()', () => {
+    suite('generateTreeItemInput()', () => {
 
-        test('no treeData return only root', () => {
+        test('no treeData return Input root with Model and Request', () => {
             const treeData = {};
             const formatData: FormatModel = {
-                format: 1,
+                format: Format.json,
                 formattedInput: '',
                 formattedJson: JSON.stringify(treeData),
             };
-            const beginPath = 'Model';
-            const result = TreeBuilder.generateTreeItemModel(formatData, beginPath);
+            const detectedModel: DetectedModel = {
+                model: formatData
+            };
+            const result = TreeBuilder.generateTreeItemInput(detectedModel);
 
-            chai.expect(result.name).to.equal(beginPath);
-            chai.expect(result.path).to.equal(beginPath);
-            chai.expect(result.children).undefined;
+            chai.expect(result.name).to.equal('Input');
+            chai.expect(result.path).to.equal('');
             chai.expect(result.isCollection).undefined;
             chai.expect(result.exampleValue).undefined;
 
+            const children = result.children;
+            if (!children){
+                fail('Input should have children');
+            }
+            chai.expect(children.length).equal(2);
+            chai.expect(children[0].name).equal('Model');
+            chai.expect(children[0].path).equal('Model');
+            chai.expect(children[0].children).undefined;
+
+            chai.expect(children[1].name).equal('Request');
+            chai.expect(children[1].path).equal('Request');
+            const reqChildren = children[1].children;
+            if (!reqChildren)
+            {
+                fail('Request should have children');
+            }
+            
+            chai.expect(reqChildren.length).equal(2);
+            chai.expect(reqChildren[0]).deep.equal({ name: 'Method', path: 'Request.Method', exampleValue: 'null'});
+            chai.expect(reqChildren[1]).deep.equal({ name: 'Headers', path: 'Request.Headers', exampleValue: 'null'});
         });
 
-        test('simple property types, can be handled', () => {
+        test('Request: taken from input.httpRequest', () => {
+
+            /* eslint-disable @typescript-eslint/naming-convention */
+            const detectedModel: DetectedModel = {
+                model: {
+                    format: Format.json,
+                    formattedInput: '',
+                    formattedJson: JSON.stringify({})
+                },
+                httpRequest: {
+                    method: 'POST',
+                    headers: {
+                        'Content-Type': 'application/json',
+                        'Custom': 'testing'
+                    }
+                }
+            };
+            /* eslint-enable @typescript-eslint/naming-convention */
+
+            const result = TreeBuilder.generateTreeItemInput(detectedModel);
+            const children = result.children;
+            if (!children){
+                fail('Input should have children');
+            }
+            chai.expect(children.length).equal(2);
+
+            chai.expect(children[1].name).equal('Request');
+            chai.expect(children[1].path).equal('Request');
+            const reqChildren = children[1].children;
+            if (!reqChildren)
+            {
+                fail('Request should have children');
+            }
+            
+            chai.expect(reqChildren.length).equal(2);
+            chai.expect(reqChildren[0]).deep.equal({ name: 'Method', path: 'Request.Method', exampleValue: 'POST'});
+
+            chai.expect(reqChildren[1].name).equal('Headers');
+            chai.expect(reqChildren[1].path).equal('Request.Headers');
+            const headerChildren = reqChildren[1].children;
+            if (!headerChildren){
+                fail('Headers should have children');
+            }
+            chai.expect(headerChildren.length).equal(2);
+            chai.expect(headerChildren[0]).deep.equal({ name: 'Content-Type', path: "Request.Headers['Content-Type']", exampleValue: 'application/json'});
+            chai.expect(headerChildren[1]).deep.equal({ name: 'Custom', path: 'Request.Headers.Custom', exampleValue: 'testing'});
+        });
+
+        test('Model: simple property types, can be handled', () => {
             const treeData = {
                 numeric: 123,
                 text: 'abc',
                 nothing: null
             };
-            const formatData: FormatModel = {
-                format: 1,
-                formattedInput: '',
-                formattedJson: JSON.stringify(treeData),
-            };
-            const beginPath = 'Model';
-            const result = TreeBuilder.generateTreeItemModel(formatData, beginPath);
 
-            chai.expect(result.isCollection).undefined;
-            chai.expect(result.exampleValue).undefined;
-            const children = result.children;
-            if (!children) {
-                fail();
-            }
+            const children = _getModelChildren(treeData);
+            
             chai.expect(children.length).equal(3);
             chai.expect(children[0]).deep.equal({ name: 'numeric', path: 'Model.numeric', exampleValue: '123' });
             chai.expect(children[1]).deep.equal({ name: 'text', path: 'Model.text', exampleValue: 'abc' });
             chai.expect(children[2]).deep.equal({ name: 'nothing', path: 'Model.nothing', exampleValue: 'null' });
         });
 
-        test('string array, adds items', () => {
+        test('Model: string array, adds items', () => {
             const treeData = {
                 texts: [
                     'option1',
                     'option2'
                 ]
             };
-            const formatData: FormatModel = {
-                format: 1,
-                formattedInput: '',
-                formattedJson: JSON.stringify(treeData),
-            };
-            const beginPath = 'Model';
-            const result = TreeBuilder.generateTreeItemModel(formatData, beginPath);
+            
+            const modelChildren = _getModelChildren(treeData);
 
-            if (!result.children) {
-                fail();
-            }
-            const textChild = result.children[0];
+            chai.expect(modelChildren.length).equal(1);
+            const textChild = modelChildren[0];
 
             chai.expect(textChild.name).equal('texts');
             chai.expect(textChild.path).equal('Model.texts');
@@ -86,7 +137,7 @@ suite('TreeBuilder Tests', () => {
             chai.expect(children[1]).deep.equal({ name: '1', path: 'Model.texts[1]', exampleValue: 'option2' });
         });
 
-        test('object array, adds properties of first item', () => {
+        test('Model: object array, adds properties of first item', () => {
             const treeData = {
                 objects: [
                     { simple: 'simple1', prop: 'prop1', other: 'a' },
@@ -94,22 +145,14 @@ suite('TreeBuilder Tests', () => {
                     { simple: 'simple3', prop: 'prop3', other: 'c' },
                 ]
             };
-            const formatData: FormatModel = {
-                format: 1,
-                formattedInput: '',
-                formattedJson: JSON.stringify(treeData),
-            };
-            const beginPath = 'Model';
-            const result = TreeBuilder.generateTreeItemModel(formatData, beginPath);
+            const modelChildren = _getModelChildren(treeData);
 
-            if (!result.children) {
-                fail();
-            }
-            const textChild = result.children[0];
-            chai.expect(textChild.isCollection).true;
-            chai.expect(textChild.name).equal('objects');
-            chai.expect(textChild.path).equal('Model.objects');
-            const children = textChild.children;
+            chai.expect(modelChildren.length).equal(1);
+            const objChild = modelChildren[0];
+            chai.expect(objChild.name).equal('objects');
+            chai.expect(objChild.path).equal('Model.objects');
+
+            const children = objChild.children;
             if (!children) {
                 fail();
             }
@@ -119,6 +162,38 @@ suite('TreeBuilder Tests', () => {
             chai.expect(children[1]).deep.equal({ name: 'prop', path: 'x.prop', exampleValue: 'prop1' });
             chai.expect(children[2]).deep.equal({ name: 'other', path: 'x.other', exampleValue: 'a' });
         });
+
+        // Helpers
+
+        function _getModelChildren(treeData: unknown): TreeItem[] {
+            const formatData: FormatModel = {
+                format: Format.json,
+                formattedInput: '',
+                formattedJson: JSON.stringify(treeData),
+            };
+            const detectedModel: DetectedModel = {
+                model: formatData
+            };
+            const result = TreeBuilder.generateTreeItemInput(detectedModel);
+
+            chai.expect(result.isCollection).undefined;
+            chai.expect(result.exampleValue).undefined;
+
+            const inputChildren = result.children;
+            if (!inputChildren){
+                fail('Input should have children');
+            }
+            const model = inputChildren[0];
+            if (!model) {
+                fail();
+            }
+
+            const children = model.children;
+            if (!children){
+                fail('Model should have children');
+            }
+            return children;
+        } 
     });
 
     suite('generateTreeItemStep()', () => {
@@ -175,16 +250,19 @@ suite('TreeBuilder Tests', () => {
             const stepType = CONSTANTS.httpStepConfigurationType;
             
             const result = TreeBuilder.generateTreeItemStep(stepId, stepType, { 
-                format: Format.json, formattedInput: '',
-                formattedJson: `{
-                    "testProp1": "value1",
-                    "testObj1": {
-                        "testSubProp1": "value2"
-                    },
-                    "testList1": [
-                        "value3"
-                    ]
-                }`
+                model: {
+                    format: Format.json, 
+                    formattedInput: '',
+                    formattedJson: `{
+                        "testProp1": "value1",
+                        "testObj1": {
+                            "testSubProp1": "value2"
+                        },
+                        "testList1": [
+                            "value3"
+                        ]
+                    }`
+                }
             });
 
             if (!result.children) {